Rename the files and extract some logic to ext
