+++ pkg/sdk/roles_impl_gen.go
@@ -58,15 +58,3 @@ func (v *roles) Grant(ctx context.Context, req *GrantRoleRequest) error {
 func (v *roles) Revoke(ctx context.Context, req *RevokeRoleRequest) error {
 	return validateAndExec(v.client, ctx, req.toOpts())
 }
-
-func (v *roles) RevokeSafely(ctx context.Context, req *RevokeRoleRequest) error {
-	return SafeRevokePrivileges(func() error { return v.Revoke(ctx, req) })
-}
-
-func (v *roles) Use(ctx context.Context, req *UseRoleRequest) error {
-	return v.client.Sessions.UseRole(ctx, req.id)
-}
-
-func (v *roles) UseSecondary(ctx context.Context, req *UseSecondaryRolesRequest) error {
-	return v.client.Sessions.UseSecondaryRoles(ctx, req.option)
-}
